test: refactor desugar test
It should test the directive directly instead of going through the
omnifset parsing as well.
